pyproject.toml
[tool.poetry]
name = "mpscenes"
version = "0.4.1"
description = "Generic motion planning scenes, including goals and obstacles."
authors = ["Max <m.spahn@tudelft.nl>"]
  

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22.0"
PyYAML = "^6.0"
geomdl = "^5.3.1"
pyquaternion = "^0.9.9"
omegaconf = "^2.2.2"
pybullet = {version = "^3.2.3", optional = true}
matplotlib = "^3.3.0"
sympy = "^1.7"

[tool.poetry.extras]
bullet = ["pybullet"]

[tool.poetry.dev-dependencies]
pylint = "^2.12"
pytest = "^6.2.5"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
